+++ CHANGELOG.md
@@ -59,8 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Debugger: Prevent stack overflows when expanding "static" section in probe-rs-debugger. (#1231)
 - RTT: Prevent panicking in `probe-rs-cli-util/src/rtt/rs` when defmt stream decoding provides invalid frame index. (#1236)
 - Fix: Attaching to LPC55S69 seems to stop code execution - incorrect values in target YAML. (#1220)
-- Fix: Panic during flashing, because targets/STM32H7.yaml has invalid core name for dual-core devices. (#1023)
-- Fix: STM32H7 not all ram sections are described (#429).
+- Debug: Fix `probe-rs-debugger` crashes when variable unwind fails with excessively long error messages. (#1252)
 
 ## [0.13.0]